Extract & add search functionality to cardgrid
